requirements.txt
flask>=2.2.5
mysql-connector-python==8.0.21